exceptions: move WazoConfdUnreachable to plugin_helpers
Why:

* comply with wazo-notebook/file-structure.md
