+++ wazo_calld/exceptions.py
@@ -91,35 +91,3 @@ class TokenWithUserUUIDRequiredError(APIException):
             message='A valid token with a user UUID is required',
             error_id='token-with-user-uuid-required',
         )
-
-
-class WazoConfdError(APIException):
-
-    def __init__(self, confd_client, error):
-        super().__init__(
-            status_code=503,
-            message='wazo-confd error',
-            error_id='wazo-confd-error',
-            details={
-                'wazo_confd_config': {'host': confd_client.host,
-                                      'port': confd_client.port,
-                                      'timeout': confd_client.timeout},
-                'original_error': str(error),
-            }
-        )
-
-
-class WazoConfdUnreachable(APIException):
-
-    def __init__(self, confd_client, error):
-        super().__init__(
-            status_code=503,
-            message='wazo-confd server unreachable',
-            error_id='wazo-confd-unreachable',
-            details={
-                'wazo_confd_config': {'host': confd_client.host,
-                                      'port': confd_client.port,
-                                      'timeout': confd_client.timeout},
-                'original_error': str(error),
-            }
-        )
