+++ wazo_calld/plugin_helpers/exceptions.py
@@ -110,3 +110,19 @@ class NoSuchVoicemail(APIException):
                 'voicemail_id': voicemail_id,
             }
         )
+
+
+class WazoConfdUnreachable(APIException):
+
+    def __init__(self, confd_client, error):
+        super().__init__(
+            status_code=503,
+            message='wazo-confd server unreachable',
+            error_id='wazo-confd-unreachable',
+            details={
+                'wazo_confd_config': {'host': confd_client.host,
+                                      'port': confd_client.port,
+                                      'timeout': confd_client.timeout},
+                'original_error': str(error),
+            }
+        )
